refactor(frontend): decompose HomeDashboardParts into modular files
Separated the "Mega Widget" `HomeDashboardParts.tsx` into dedicated components according to the Single Responsibility Principle:
- Extracted `HomeSurfaceCard` and `HomeActionTile` to `HomeCard.tsx`
- Extracted `HomeChatRow` to `HomeRecentChatRow.tsx`
- Extracted `HomeTaskRow` to `HomeTaskRow.tsx`
- Extracted `HomeAgentBadge` to `HomeAgentBadge.tsx`
- Extracted `HomeHeroCard` to `HomeHeroCard.tsx`
- Deleted zombie code `HomeAgentChip.tsx`
- Updated imports in `home.tsx` to reference the newly separated modules.
- Refactored `index.ts` to export the new modules cleanly.
