+++ frontend/src/components/home/HomeSectionHeader.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { View } from 'react-native';
-import { useColorScheme } from 'nativewind';
-import { AppText } from '../AppText';
-import { ScalePressable } from '../ScalePressable';
+import { AppText } from '@/components/AppText';
+import { ScalePressable } from '@/components/ScalePressable';
+import { HOME_COLORS } from './homeTheme';
 
 export function HomeSectionHeader({
   title,
@@ -13,21 +13,25 @@ export function HomeSectionHeader({
   actionLabel?: string;
   onAction?: () => void;
 }) {
-  const { colorScheme } = useColorScheme();
-  const isDark = colorScheme === 'dark';
-
   return (
-    <View className="flex-row justify-between items-center mb-6">
-      <AppText variant="h3" className={isDark ? 'text-onSurface-dark' : 'text-onSurface-light'}>
+    <View
+      style={{
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        marginBottom: 12,
+      }}
+    >
+      <AppText variant="h3" style={{ color: HOME_COLORS.text, fontWeight: '700' }}>
         {title}
       </AppText>
-      {actionLabel && onAction && (
+      {actionLabel && onAction ? (
         <ScalePressable onPress={onAction}>
-          <AppText variant="bodySm" className="text-primary-DEFAULT">
+          <AppText variant="bodySm" style={{ color: HOME_COLORS.primary, fontWeight: '700' }}>
             {actionLabel}
           </AppText>
         </ScalePressable>
-      )}
+      ) : null}
     </View>
   );
 }
