+++ frontend/src/components/home/index.ts
@@ -3,5 +3,6 @@ export * from './HomeSectionHeader';
 export * from './HomeQuickAction';
 export * from './HomeRecentChatRow';
 export * from './HomeTaskRow';
-export * from './HomeAgentChip';
+export * from './HomeAgentBadge';
 export * from './HomeNewChatModal';
+export * from './HomeHeroCard';
